requirements.txt
h5py
pyproj
pandas
matplotlib
pyarrow
segmentation_models_pytorch
lightning
datasets
"huggingface_hub[cli]"
tensorflow
timm